chore: add `fedimint-empty` module
Initiaially the goal of `fedimint-dummy` module was to be a good template to start a new module.

But with time, it was hijacked to be become `fedimint-`"print me money out of thin air" module (aka `fedimint-federal-reserve` module),
and it now actually contains quite a bit of code.

So when trying to use it as a starting point, one needs to go a rename,
delete and fix more code than neccessary.

This module ius a clone of `fedimint-dummy`, but with almost everything
removed.