chore(client): retry api version discovery for at least 30 seconds
During debugging with joschisan, we've spotted a case where in a test
the Federation was not yet fully up (peer api servers were down), yet
the test Client was already attempting to join with a pre-prepared
config.

Normally in this case fetching the config would fail, but that
was passed straight from the server fixture, so api discovery
call was actually the first call. It queried all peers, all were
down, the request logic would re-attempt connection but only once,
which instantly failed as well, failing the whole initial api
version discovery and disabling all client modules.

While in this case it seems like a combination of test failing to
setup Federation before attempting to run client code, it makes
me think that in real life there might be cases where the client
is joining new Federation, gets a config, and then e.g. switches
from wifi to GSM, etc. and has a brief period of being offline.

Again - this is only ever a problem on a first time client is
joining. But to prevent such a hiccups, in this change,
I'm making it so the client will attempt at least for a certain
period to discover the version before giving it up, to lower
the chances of such an unfortunate UX.